% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastSplit.R
\name{fastSplit}
\alias{fastSplit}
\title{Splits the input string using the delimiter character.}
\usage{
fastSplit(s, delim)
}
\arguments{
\item{s}{A string}

\item{delim}{A delimiter string, which should be one character.}
}
\value{
A character vecter
}
\description{
Splits the input string using the delimiter character.
}
